requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.3
python-dotenv==1.0.1
supabase==2.1.0
joblib==1.4.2
scikit-learn==1.5.0
PyJWT==2.9.0